move exhibitions into getter/parser
