+++ server/services/prismic.js
@@ -1,5 +1,9 @@
 import {prismicApi, prismicPreviewApi} from './prismic-api';
-import {parseArticleDoc, parseEventDoc, parseWebcomicDoc} from './prismic-parsers';
+import {
+  parseArticleDoc, parseEventDoc, parseExhibitionsDoc, parsePromoListItem,
+  parseWebcomicDoc
+} from './prismic-parsers';
+import type {Promo} from '../model/promo';
 
 type DocumentType = 'articles' | 'webcomics' | 'events' | 'exhibitions';
 
@@ -37,9 +41,48 @@ export async function getArticle(id: string, previewReq: ?Request) {
 
 export async function getEvent(id: string, previewReq: ?Request): Promise<?Event> {
   const fetchLinks = eventFields.concat(peopleFields, contributorFields);
-  const event = await getTypeById(previewReq, ['event'], id, {fetchLinks});
+  const event = await getTypeById(previewReq, ['events'], id, {fetchLinks});
 
   if (!event) { return null; }
 
   return parseEventDoc(event);
+}
+
+
+type ExhibitionContent = {|
+  exhibition: Exhibition;
+  galleryLevel: string;
+  relatedBooks: Array<Promo>;
+  relatedEvents: Array<Promo>;
+  relatedGalleries: Array<Promo>;
+  relatedArticles: Array<Promo>;
+  imageGallery: any;
+  textAndCaptionsDocument: any;
+|}
+
+export async function getExhibition(id: string, previewReq: ?Request): Promise<?ExhibitionContent> {
+  const exhibition = await getTypeById(previewReq, ['exhibitions'], id, {});
+
+  if (!exhibition) { return null }
+
+  const ex = parseExhibitionsDoc(exhibition);
+
+  const promoList = exhibition.data.promoList;
+  const relatedArticles = promoList.filter(x => x.type === 'article').map(parsePromoListItem);
+  const relatedEvents = promoList.filter(x => x.type === 'event').map(parsePromoListItem);
+  const relatedBooks = promoList.filter(x => x.type === 'book').map(parsePromoListItem);
+  const relatedGalleries = promoList.filter(x => x.type === 'gallery').map(parsePromoListItem);
+
+  const sizeInKb = Math.round(exhibition.data.textAndCaptionsDocument.size / 1024);
+  const textAndCaptionsDocument = Object.assign({}, exhibition.data.textAndCaptionsDocument, {sizeInKb});
+
+  return {
+    exhibition: ex,
+    galleryLevel: '0',
+    textAndCaptionsDocument: textAndCaptionsDocument.url && textAndCaptionsDocument,
+    relatedBooks: relatedBooks,
+    relatedEvents: relatedEvents,
+    relatedGalleries: relatedGalleries,
+    relatedArticles: relatedArticles
+  };
 }
